The merge resolution carried this branch's stale dependency pins over
upstream's newer ones, silently downgrading thiserror to 2.0.18 and
tokio to 1.52.3 while Cargo.lock had already resolved to upstream's
2.0.19 and 1.53.1. Restored both so manifest and lockfile agree.

Also moved iggy_connector_mysql_source into its sorted position in the
connector_plugins default list rather than leaving it appended.

+++ foreign/go/contracts/meta_data.go
@@ -76,10 +76,10 @@ func (m *ClusterMetadata) UnmarshalBinary(data []byte) error {
 	if err := binary.Read(r, binary.LittleEndian, &nameLen); err != nil {
 		return err
 	}
-	if int(nameLen) > r.Len() {
+	if uint64(nameLen) > uint64(r.Len()) {
 		return errors.New("invalid name length")
 	}
-	nameb := make([]byte, nameLen)
+	nameb := make([]byte, int(nameLen))
 	if _, err := r.Read(nameb); err != nil {
 		return err
 	}
